BUG/MEDIUM: servers: Add a per-thread counter of idle connections.
Add a per-thread counter of idling connections, and use it to determine
how many connections we should kill after the timeout, instead of using
the global counter, or we're likely to just kill most of the connections.

This should be backported to 1.9.

(cherry picked from commit f131481a0af79037bc6616edf450ae81d80084d7)
Signed-off-by: Willy Tarreau <w@1wt.eu>