BUG/MEDIUM: cache: Get objects from the cache only for GET and HEAD requests
Only responses for GET requests are stored in the cache. But there is no check
on the method during the lookup. So it is possible to retrieve an object from
the cache independently of the method, from the time the key of the object
matches. Now, lookups are performed only for GET and HEAD requests.

This patch must be backportedi in 1.9.

(cherry picked from commit b3d4bca415ab90108b29b232a4ae7a892d4c48ec)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>