BUG/MINOR: cache/htx: Return only the headers of cached objects to HEAD requests
The body of a cached object must not be sent in response to a HEAD request. This
works for the legacy HTTP because the parsing is performed by HTTP analyzers
_AND_ because the connection is closed at the end of the transaction. So the
body is ignored. But the applet send it. For the HTX, the applet must skip the
body explicitly.

This patch must be backported to 1.9.

(cherry picked from commit f0dd03745647f91bb4adba1412e51a828cd112bc)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>