BUG/MINOR: listener: keep accept rate counters accurate under saturation
The test on l->nbconn forces to exit the loop before updating the freq
counters, so the last session which reaches a listener's limit will not
be accounted for in the session rate measurement.

Let's move the test at the beginning of the loop and mark the listener
as saturated on exit.

This may be backported to 1.9 and 1.8.

(cherry picked from commit 741b4d6b7aad1e4a66dd8584b5eff729b08fade7)
Signed-off-by: William Lallemand <wlallemand@haproxy.org>