BUG/MAJOR: listener: Make sure the listener exist before using it.
In listener_accept(), make sure we have a listener before attempting to
use it.
An another thread may have closed the FD meanwhile, and set fdtab[fd].owner
to NULL.
As the listener is not free'd, it is ok to attempt to accept() a new
connection even if the listener was closed. At worst the fd has been
reassigned to another connection, and accept() will fail anyway.

Many thanks to Richard Russo for reporting the problem, and suggesting the
fix.

This should be backported to 1.9 and 1.8.

(cherry picked from commit d16a9dfed80e75d730754b717370515265698cdd)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit a2511ed1fcdfa8047dbe2268fc0259f9b06cf891)
Signed-off-by: William Lallemand <wlallemand@haproxy.org>