BUG/MEDIUM: h2: advertise to servers that we don't support push
The h2c_send_settings() function was initially made to serve on the
frontend. Here we don't need to advertise that we don't support PUSH
since we don't do that ourselves. But on the backend side it's
different because PUSH is enabled by default so we must announce that
we don't want the server to use it.

This must be backported to 1.9.

(cherry picked from commit 0bbad6bb06f39185a1b6803e354dac1a48be6c16)
Signed-off-by: Willy Tarreau <w@1wt.eu>