BUG/MAJOR: stats: Fix how huge POST data are read from the channel
When the body length is greater than a chunk size (so if length of POST data
exceeds the buffer size), the requests is rejected with the status code
STAT_STATUS_EXCD. Otherwise the stats applet will wait to have all the data to
copy and parse them. But there is a problem when the total request size
(including the headers) is just lower than the buffer size but greater the
buffer size less the reserve. In such case, the body length is considered as
enough small to be processed but not entierly received. So the stats applet
waits for more data. But because outgoing data are still there, the channel's
buffer is considered as full and nothing more can be read, leading to a freeze
of the session.

Note this bug is pretty easy to reproduce with the legacy HTTP. It is harder
with the HTX but still possible. To fix the bug, in the stats applet, when the
request is not fully received, we check if at least the reserve remains
available the channel's buffer.

This patch must be backported as far as 1.5. But because the HTX does not exist
in 1.8 and lower, it will have to be adapted for these versions.

(cherry picked from commit 2f9a41d52b28b88d44aab063bb2b88025a388981)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit 18a5b1eece7e12047c1dc371c42f72d9d129ce0a)
[cf: HTX code removed and calls to the new buffer API replaced to use the old
one]
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>