MINOR: tools: implement my_flsl()
We already have my_ffsl() to find the lowest bit set in a word, and
this patch implements the search for the highest bit set in a word.
On x86 it uses the bsr instruction and on other architectures it
uses an efficient implementation.

(cherry picked from commit d87a67f9bc422bf41b6b81c1e99d9aebbbc18d8e)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>

[Cf: This one is required to backport some patches on the SPOE. It is just a new
function, so there is no impact]