BUG/MINOR: mux-h1: Don't report an error on EOS if no message was received
An error is reported if the EOS is detected before the end of the message. But
we must be carefull to not report an error if there is no message at all.

This patch must be backported to 1.9.

(cherry picked from commit 269223886db08d4bf8fc038e0794f181b939ce3d)
Signed-off-by: Willy Tarreau <w@1wt.eu>