BUG/MAJOR: tasks: Use the TASK_GLOBAL flag to know if we're in the global rq.
In task_unlink_rq, to decide if we should logk the global runqueue lock,
use the TASK_GLOBAL flag instead of relying on t->thread_mask being tid_bit,
as it could be so while still being in the global runqueue if another thread
woke that task for us.

This should be backported to 1.9.

(cherry picked from commit 1d7f37a2cb395323605e474c1539836e609734b0)
Signed-off-by: Olivier Houchard <cognet@ci0.org>