BUILD: Makefile: allow the reg-tests target to be verbose
When debugging reg-tests, it's quite annoying not to be able to figure
the syntax to call the scripts. Let's replace the '@' with '$(Q)' as for
other commands so that launching them with "V=1" is enough to reveal the
command line.

(cherry picked from commit f22bd9055bb49624b5ea2589fa40deb06e904af7)
Signed-off-by: Willy Tarreau <w@1wt.eu>