BUG/MINOR: http/counters: fix missing increment of fe->srv_aborts
When a server aborts a transfer, we used to increment the backend's
counter but not the frontend's during the forwarding phase. This fixes
it. It might be backported to all supported versions (possibly removing
the htx part) though it is of very low importance.

(cherry picked from commit d1fd6f5f64e4d05d4993f2d43c1ee8c79a16fec1)
[wt: s/_HA_ATOMIC/HA_ATOMIC/]
Signed-off-by: Willy Tarreau <w@1wt.eu>