BUG/MEDIUM: mux-h2: Make sure we destroyed the h2s once shutr/shutw is done.
If we're waiting until we can send a shutr and/or a shutw, once we're done
and not considering sending anything, destroy the h2s, and eventually the
h2c if we're done with the whole connection, or it will never be done.

This should be backported to 1.9.

(cherry picked from commit 7a977431ca4864d8aaa148ef29369ca7f5fcfd5e)
Signed-off-by: Willy Tarreau <w@1wt.eu>