BUG/MEDIUM: mux-h2: Don't bother keeping the h2s if detaching and nothing to send.
In h2_detach(), don't bother keeping the h2s even if it was waiting for
flow control if we no longer are subscribed for receiving or sending, as
nobody will do anything once we can write in the mux, anyway. Failing to do
so may lead to h2s being kept opened forever.

This should be backported to 1.9.

(cherry picked from commit 16ff261633ddae1b86f897ee2c14bd8fdda92401)
Signed-off-by: Willy Tarreau <w@1wt.eu>